updated qr code
