+++ HackIllinois/Assets.xcassets/QR Code Logo.imageset/Contents.json
