+++ HackIllinois/Extensions/CIImage.swift
@@ -1,0 +1,64 @@
+//
+//  CIImage.swift
+//  HackIllinois
+//
+//  Created by HackIllinois Team on 12/13/21.
+//  Copyright © 2021 HackIllinois. All rights reserved.
+//  This file is part of the Hackillinois iOS App.
+//  The Hackillinois iOS App is open source software, released under the University of
+//  Illinois/NCSA Open Source License. You should have received a copy of
+//  this license in a file with the distribution.
+//
+
+import Foundation
+import UIKit
+
+//https://www.avanderlee.com/swift/qr-code-generation-swift/
+extension CIImage {
+    /// Combines the current image with the given image centered.
+    func combined(with image: CIImage) -> CIImage? {
+        guard let combinedFilter = CIFilter(name: "CISourceOverCompositing") else { return nil }
+        let centerTransform = CGAffineTransform(translationX: extent.midX - (image.extent.size.width / 2), y: extent.midY - (image.extent.size.height / 2))
+        combinedFilter.setValue(image.transformed(by: centerTransform), forKey: "inputImage")
+        combinedFilter.setValue(self, forKey: "inputBackgroundImage")
+        return combinedFilter.outputImage!
+    }
+    /// Inverts the colors and creates a transparent image by converting the mask to alpha.
+    /// Input image should be black and white.
+    var transparent: CIImage? {
+        return inverted?.blackTransparent
+    }
+
+    /// Inverts the colors.
+    var inverted: CIImage? {
+        guard let invertedColorFilter = CIFilter(name: "CIColorInvert") else { return nil }
+
+        invertedColorFilter.setValue(self, forKey: "inputImage")
+        return invertedColorFilter.outputImage
+    }
+
+    /// Converts all black to transparent.
+    var blackTransparent: CIImage? {
+        guard let blackTransparentFilter = CIFilter(name: "CIMaskToAlpha") else { return nil }
+        blackTransparentFilter.setValue(self, forKey: "inputImage")
+        return blackTransparentFilter.outputImage
+    }
+
+    /// Applies the given color as a tint color.
+    func tinted(using color: UIColor) -> CIImage?
+    {
+        guard
+            let transparentQRImage = transparent,
+            let filter = CIFilter(name: "CIMultiplyCompositing"),
+            let colorFilter = CIFilter(name: "CIConstantColorGenerator") else { return nil }
+
+        let ciColor = CIColor(color: color)
+        colorFilter.setValue(ciColor, forKey: kCIInputColorKey)
+        let colorImage = colorFilter.outputImage
+
+        filter.setValue(colorImage, forKey: kCIInputImageKey)
+        filter.setValue(transparentQRImage, forKey: kCIInputBackgroundImageKey)
+
+        return filter.outputImage!
+    }
+}
