+++ HackIllinois/Extensions/UIImage+QRCode.swift
@@ -38,4 +38,60 @@ extension UIImage {
 
         self.init(cgImage: cgImage)
     }
+    
+    convenience init?(qrString: String, color: UIColor, backgroundImage: UIImage, logo: UIImage) {
+        // Create simple QR Code
+        guard let qrFilter = CIFilter(name: "CIQRCodeGenerator"),
+              let data = qrString.data(using: .isoLatin1, allowLossyConversion: false)
+            else { return nil }
+        
+        qrFilter.setDefaults()
+        qrFilter.setValue(data, forKey: "inputMessage")
+        
+        // Bump up error correction
+        qrFilter.setValue("H", forKey: "inputCorrectionLevel")
+        
+        let firstScaleTransform = CGAffineTransform(scaleX: 9, y: 9)
+        let initialScaledQRCode = qrFilter.outputImage?.transformed(by: firstScaleTransform)
+        
+        guard let maskToAlphaFilter = CIFilter(name: "CIMaskToAlpha") else { return nil }
+        maskToAlphaFilter.setDefaults()
+        maskToAlphaFilter.setValue(initialScaledQRCode, forKey: kCIInputImageKey)
+        
+        let simpleQRCode = maskToAlphaFilter.outputImage
+                
+        // Add background to the center of the QR Code
+        guard let backgroundCGImage = backgroundImage.cgImage,
+              let invertedQRCode = simpleQRCode?.inverted,
+              let tintedQRCode = invertedQRCode.tinted(using: color),
+              let backgroundQRCode = tintedQRCode.combined(with: CIImage(cgImage: backgroundCGImage))
+        else { return nil }
+                
+        let secondScaleTransform = CGAffineTransform(scaleX: 2.5, y: 2.5)
+        let scaledBackgroundQRCode = backgroundQRCode.transformed(by: secondScaleTransform)
+        
+        // Add logo on top of the round background
+        guard let logoCGImage = logo.cgImage,
+              let customQRCode = scaledBackgroundQRCode.combined(with: CIImage(cgImage: logoCGImage))
+            else { return nil }
+        
+        self.init(ciImage: customQRCode)
+    }
+    
+}
+
+// https://gist.github.com/brownsoo/1b772612b54c4dc58d88ae71aec19552
+public extension UIImage {
+  func round(_ radius: CGFloat) -> UIImage {
+        let rect = CGRect(x: 0, y: 0, width: size.width, height: size.height)
+        let renderer = UIGraphicsImageRenderer(size: rect.size)
+        let result = renderer.image { c in
+            let rounded = UIBezierPath(roundedRect: rect, cornerRadius: radius)
+            rounded.addClip()
+            if let cgImage = self.cgImage {
+                UIImage(cgImage: cgImage, scale: self.scale, orientation: self.imageOrientation).draw(in: rect)
+            }
+        }
+        return result
+    }
 }
